RHOAIENG-79889: Extract DashboardModalFooter, DashboardSearchField, DashboardExpandableSection to ui-core
