+++ frontend/src/concepts/pipelines/content/ArchiveModal.tsx
@@ -12,7 +12,7 @@ import {
 } from '@patternfly/react-core';
 import { TrackingOutcome } from '@odh-dashboard/ui-core';
 import { usePipelinesAPI } from '#~/concepts/pipelines/context';
-import DashboardModalFooter from '#~/concepts/dashboard/DashboardModalFooter';
+import DashboardModalFooter from '@odh-dashboard/ui-core/components/DashboardModalFooter';
 import { fireFormTrackingEvent } from '#~/concepts/analyticsTracking/segmentIOUtils';
 
 interface ArchiveModalProps {
