+++ frontend/src/pages/modelRegistrySettings/DeleteModelRegistryModal.tsx
@@ -11,7 +11,7 @@ import {
 } from '@patternfly/react-core';
 import { TrackingOutcome } from '@odh-dashboard/ui-core';
 import { ModelRegistryKind } from '#~/k8sTypes';
-import DashboardModalFooter from '#~/concepts/dashboard/DashboardModalFooter';
+import DashboardModalFooter from '@odh-dashboard/ui-core/components/DashboardModalFooter';
 import { deleteModelRegistryBackend } from '#~/services/modelRegistrySettingsService';
 import { fireFormTrackingEvent } from '#~/concepts/analyticsTracking/segmentIOUtils';
