+++ frontend/src/pages/modelServing/screens/projects/nim/NIMServiceModal/__tests__/ManageNIMServingModal.spec.tsx
@@ -151,7 +151,7 @@ jest.mock('../../../ServingRuntimeModal/AuthServingRuntimeSection', () => ({
   default: jest.fn(() => <div data-testid="auth-serving-runtime-section">Auth Section</div>),
 }));
 
-jest.mock('#~/concepts/dashboard/DashboardModalFooter', () => ({
+jest.mock('@odh-dashboard/ui-core/components/DashboardModalFooter', () => ({
   __esModule: true,
   default: jest.fn(() => <div data-testid="modal-footer">Modal Footer</div>),
 }));
