+++ frontend/src/pages/pipelines/global/runs/RestoreRunWithArchivedExperimentModal.tsx
@@ -16,7 +16,7 @@ import { usePipelinesAPI } from '#~/concepts/pipelines/context';
 import useRestoreStatuses from '#~/concepts/pipelines/content/useRestoreStatuses';
 import { ExperimentKF, PipelineRunKF } from '#~/concepts/pipelines/kfTypes';
 import { allSettledPromises } from '#~/utilities/allSettledPromises';
-import DashboardModalFooter from '#~/concepts/dashboard/DashboardModalFooter';
+import DashboardModalFooter from '@odh-dashboard/ui-core/components/DashboardModalFooter';
 import RunsWithArchivedExperimentTable from './RunsWithArchivedExperimentTable';
 import { PipelineRunTabTitle } from './types';
