+++ frontend/src/pages/projects/screens/spawner/storage/AttachExistingStorageModal.tsx
@@ -9,7 +9,7 @@ import {
 } from '@patternfly/react-core';
 import React from 'react';
 import { ExistingStorageObject, MountPath } from '#~/pages/projects/types';
-import DashboardModalFooter from '#~/concepts/dashboard/DashboardModalFooter';
+import DashboardModalFooter from '@odh-dashboard/ui-core/components/DashboardModalFooter';
 import SpawnerMountPathField from './SpawnerMountPathField';
 import AddExistingStorageField from './AddExistingStorageField';
 import { MOUNT_PATH_PREFIX } from './const';
