+++ packages/hardware-profiles/src/pages/nodeSelector/ManageNodeSelectorModal.tsx
@@ -8,7 +8,7 @@ import {
   ModalHeader,
   ModalFooter,
 } from '@patternfly/react-core';
-import DashboardModalFooter from '@odh-dashboard/internal/concepts/dashboard/DashboardModalFooter';
+import DashboardModalFooter from '@odh-dashboard/ui-core/components/DashboardModalFooter';
 import useGenericObjectState from '@odh-dashboard/ui-core/utilities/useGenericObjectState';
 import { useValidation } from '@odh-dashboard/ui-core/utilities/useValidation';
 import { EMPTY_NODE_SELECTOR, NodeSelectorRow } from './const';
