+++ packages/model-serving/src/components/metrics/bias/BiasConfigurationPage/BiasConfigurationModal/ManageBiasConfigurationModal.tsx
@@ -12,7 +12,7 @@ import DashboardHelpTooltip from '@odh-dashboard/ui-core/components/DashboardHel
 import { BiasMetricConfig } from '@odh-dashboard/internal/concepts/trustyai/types';
 import { BiasMetricType } from '@odh-dashboard/internal/api';
 import { TrustyAIContext } from '@odh-dashboard/internal/concepts/trustyai/context/TrustyAIContext';
-import DashboardModalFooter from '@odh-dashboard/internal/concepts/dashboard/DashboardModalFooter';
+import DashboardModalFooter from '@odh-dashboard/ui-core/components/DashboardModalFooter';
 import type { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
 import useBiasConfigurationObject from './useBiasConfigurationObject';
 import MetricTypeField from './MetricTypeField';
