+++ packages/model-serving/src/components/metrics/bias/BiasTab.tsx
@@ -14,7 +14,7 @@ import {
 import { ExclamationCircleIcon } from '@patternfly/react-icons';
 import MetricsPageToolbar from '@odh-dashboard/internal/concepts/metrics/MetricsPageToolbar';
 import { useModelBiasData } from '@odh-dashboard/internal/concepts/trustyai/context/useModelBiasData';
-import DashboardExpandableSection from '@odh-dashboard/internal/concepts/dashboard/DashboardExpandableSection';
+import DashboardExpandableSection from '@odh-dashboard/ui-core/components/DashboardExpandableSection';
 import { TrustyInstallState } from '@odh-dashboard/internal/concepts/trustyai/types';
 import BiasMetricConfigSelector from './BiasMetricConfigSelector';
 import BiasChart from './BiasChart';
