+++ packages/ui-core/package.json
@@ -71,7 +71,11 @@
     "./components/projectSelector/ProjectSelector": "./src/components/projectSelector/ProjectSelector.tsx",
     "./components/projectSelector/ProjectSelectorNavigator": "./src/components/projectSelector/ProjectSelectorNavigator.tsx",
     "./components/projectSelector/ProjectNavigatorLink": "./src/components/projectSelector/ProjectNavigatorLink.tsx",
-    "./components/projectSelector/ProjectIconWithSize": "./src/components/projectSelector/ProjectIconWithSize.tsx"
+    "./components/projectSelector/ProjectIconWithSize": "./src/components/projectSelector/ProjectIconWithSize.tsx",
+    "./components/DashboardModalFooter": "./src/components/DashboardModalFooter.tsx",
+    "./components/DashboardSearchField": "./src/components/DashboardSearchField.tsx",
+    "./components/DashboardExpandableSection": "./src/components/DashboardExpandableSection.tsx",
+    "./components/DashboardExpandableSection.scss": "./src/components/DashboardExpandableSection.scss"
   },
   "scripts": {
     "lint": "eslint .",
