+++ packages/ui-core/src/components/DashboardExpandableSection.scss
